@@ -2,9 +2,12 @@ namespace DonTasker.Interfaces
 {
     public interface ITaskContainer
     {
-        IEnumerable<ITask> Tasks { get; set; }
+        List<ITask> Tasks { get; set; }
 
         void AddTask(ITask task);
         void RemoveTask(ITask task);
+
+        //indexer
+        ITask this[int index] { get; set; }
     }
 }
